+++ llama-cli/src/mode/mod.rs
@@ -1,0 +1,2 @@
+pub mod interactive;
+pub mod repl;
